Help Modal Refactor (#483)
Wanted to get this one done in one go, but it looks like we're missing some assets. 
We'll want another go at this so holding off on the final round of crayoning until we 
have that.

Primary changes here is in formally structuring the inputs to the individual changes
to reference static assets and standardize formats. The plan is to essentially use this 
structure as a lightweight templating engine that will allow us to specify image-embeds
in the middle of a page, between paragraphs (will be v useful for when we introduce
more calculation-heavy help pages such as Liquidations).

<img width="703" alt="Screen Shot 2023-12-26 at 11 36 23 AM" src="https://github.com/Asphodel-OS/kamigotchi/assets/109483360/88a76e34-6420-4528-ac22-c0406223e837">