operator cache, LibSafeQuery (#593)
Added a `LibSafeQuery`, which allows for some limited querying that doesnt scale with `O(n)`. Implemented this in LibAccount.

Added a `CacheOperatorComponent`. While `LibSafeQuery` does introduce an `O(1)` option for querying account for operator, opted to still use a cache component instead. A `LibSafeQuery` call uses about 10k gas vs a cached component call of around 3k gas. Figured that this is worthwhile because of this bit of code is in virtually every system