Id__Comp bares (#703)
updates most ID comps to `bareComps`. These are especially impactful because its written everywhere

Notable: added a `IdPointer` to act as sort of a catch-all to the hashed pointer pattern. Its the successor to `IdOwnsCondition` but named more generally

Further: not satisfied with the current shape of `RoomGates`. Left it as such for now, but the component name for that particular application isnt great