temp 0-pruning fix: prune everything on FE (#846)
addresses the preceding 0-pruning behaviour in RECS. Does this by *forcing* every entityID on FE to be pruned. A horrible solution from a motherless developer.

Snapshot seems to store and deliver pruned IDs. Fixing that rn would be decently heavy lift, and this is a quite impactful bug because of DeterministicIDs - caused 3 accounts out of 888 could not detect their reputation. This is likely the cause of many persistent issues for a minority of players

Snapshot long overdue for another upgrade. Ideally, it shouldn't prune IDs. Until then, this maybe a stopgap